+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ### Changed
 - update README
 - Updated files: README.md
+## [0.2.8] - 2025-08-31
+
+### Changed
+- add emoji into admixture module
+- Updated files: biopytools/admixture,biopytools/cli/commands/admixture.py
+
+
+
+
 ## [0.2.7] - 2025-08-31
 
 ### Changed